Moving improved IDataItem to Expando conversion to shared extension and using it from MongoDb extension to support nesting
